requirements.txt
pbr>=1.1.0

PyYAML
cherrypy
routes
requests
openstacksdk
